oracle adapter

+++ test/helpers/DatabaseLoader.php
@@ -23,6 +23,9 @@ class DatabaseLoader
 	{
 		foreach ($this->get_fixture_tables() as $table)
 		{
+			if ($this->db->protocol == 'oci' && $table == 'rm-bldg')
+				continue;
+
 			$this->db->query('DELETE FROM ' . $this->quote_name($table));
 			$this->load_fixture_data($table);
 		}
@@ -41,7 +44,12 @@ class DatabaseLoader
 		foreach ($this->get_fixture_tables() as $table)
 		{
 			if ($this->db->protocol == 'oci')
+			{
 				$table = strtoupper($table);
+
+				if ($table == 'RM-BLDG')
+					continue;
+			}
 
 			if (in_array($table,$tables))
 				$this->db->query('DROP TABLE ' . $this->quote_name($table));
